Expand and clarify Code of Conduct document
The title "Code of Conduct" was changed to "CODE OF CONDUCT". The document was significantly expanded to include a detailed introduction, clarifying that the "Code of Conduct" should be treated like an "End User License Agreement" (EULA). The introduction specifies that the License is part of this agreement and can be found in the LICENSE.md file. It clarifies the relationship between the user and Olanorw aka Olav Sharma, explaining that "Olanorw" and "Olav Sharma" refer to the same individual.

A new section titled "Agreement" was added, outlining that by installing, copying, or using the software, the user agrees to be bound by the terms of this Agreement, ElevenLabs policies, Google Cloud policies, and the LICENSE.md file. It also mentions that even if the user opts out of seeing the agreement again by selecting "Don't show again?" when the program is launched, they are still bound by the Code of Conduct.